move reviewList and review to different components
